Added a basic example of sessions based on Cookies
