+++ vertx-web-session-stores/vertx-web-sstore-cookie/src/main/resources/META-INF/services/io.vertx.ext.web.sstore.SessionStore
@@ -1,1 +1,1 @@
-io.vertx.ext.web.sstore.cookie.CookieSessionStore
+io.vertx.ext.web.sstore.cookie.impl.CookieSessionStoreImpl
